Updated contentful client options types, refactored tests
